rearrange type converters to utils.py to prevent circular imports.
