+++ cyclopts/parameter.py
@@ -8,13 +8,11 @@ from cyclopts._convert import (
     AnnotatedType,
     convert,
     get_origin_and_validate,
-    optional_to_tuple_converter,
     resolve,
     resolve_optional,
-    to_tuple_converter,
 )
 from cyclopts.group import Group
-from cyclopts.utils import record_init
+from cyclopts.utils import optional_to_tuple_converter, record_init, to_tuple_converter
 
 
 def _double_hyphen_validator(instance, attribute, values):
